feat: Add database connection functionality
This commit adds a new file `db/conn.js` which contains the code for establishing a connection to the database using Mongoose. The `dbConnection` function is responsible for connecting to the database using the provided Atlas URI from the environment variables. If the connection is successful, a message is logged to the console. If there is an error, an error message is logged.

This commit also includes the necessary import statements for `mongoose` and `dotenv`, as well as the call to `config()` to load the environment variables.

The purpose of this commit is to enable the application to connect to the database and perform database operations.